+++ SmallBeachFarm/ModEntry.cs
@@ -27,8 +27,15 @@ namespace Pathoschild.Stardew.SmallBeachFarm
         /// <summary>The relative path to the folder containing tilesheet variants.</summary>
         private readonly string TilesheetsPath = Path.Combine("assets", "tilesheets");
 
+        /// <summary>The relative path to the folder containing tilesheet overlays.</summary>
+        private readonly string OverlaysPath = Path.Combine("assets", "overlays");
+
         /// <summary>The mod configuration.</summary>
         private ModConfig Config;
+
+        /// <summary>The minimum value to consider non-transparent.</summary>
+        /// <remarks>On Linux/Mac, fully transparent pixels may have an alpha up to 4 for some reason.</remarks>
+        private const byte MinOpacity = 5;
 
         /// <summary>A fake asset key prefix from which to load tilesheets.</summary>
         private string FakeAssetPrefix => Path.Combine("Mods", this.ModManifest.UniqueID);
@@ -100,6 +107,30 @@ namespace Pathoschild.Stardew.SmallBeachFarm
 
                 // load asset
                 Texture2D tilesheet = this.Helper.Content.Load<Texture2D>(relativePath);
+                var tilesheetPixels = new Lazy<Color[]>(() => this.GetPixels(tilesheet));
+
+                // apply overlays
+                foreach (DirectoryInfo folder in new DirectoryInfo(this.GetFullPath(this.OverlaysPath)).EnumerateDirectories())
+                {
+                    if (!this.Helper.ModRegistry.IsLoaded(folder.Name))
+                        continue;
+
+                    // get overlay
+                    Texture2D overlay = this.Helper.Content.Load<Texture2D>(Path.Combine(this.OverlaysPath, folder.Name, filename));
+                    Color[] overlayPixels = this.GetPixels(overlay);
+
+                    // apply
+                    Color[] target = tilesheetPixels.Value;
+                    for (int i = 0; i < overlayPixels.Length; i++)
+                    {
+                        Color pixel = overlayPixels[i];
+                        if (pixel.A >= ModEntry.MinOpacity)
+                            target[i] = overlayPixels[i];
+                    }
+                }
+
+                if (tilesheetPixels.IsValueCreated)
+                    tilesheet.SetData(tilesheetPixels.Value);
 
                 return (T)(object)tilesheet;
             }
@@ -149,6 +180,15 @@ namespace Pathoschild.Stardew.SmallBeachFarm
         private string GetFullPath(string relative)
         {
             return Path.Combine(this.Helper.DirectoryPath, relative);
+        }
+
+        /// <summary>Get the pixel data for a texture.</summary>
+        /// <param name="texture">The texture asset.</param>
+        private Color[] GetPixels(Texture2D texture)
+        {
+            Color[] pixels = new Color[texture.Width * texture.Height];
+            texture.GetData(pixels);
+            return pixels;
         }
 
         /// <summary>Get whether the given location is the Small Beach Farm.</summary>