+++ SmallBeachFarm/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Added support for other mods patching the custom tilesheet.
 * Added support for overlay compatibility files.
+* Added compatibility with Better Water (thanks to laulajatar!).
 * Fixed missing warp tile.
 
 ## 1.4.3